+++ src/adapter/zboss/adapter/zbossAdapter.ts
@@ -165,7 +165,7 @@ export class ZBOSSAdapter extends Adapter {
 
             return {
                 panID,
-                extendedPanID: parseInt(Buffer.from(extendedPanID).toString('hex'), 16),
+                extendedPanID: ZSpec.Utils.eui64LEBufferToHex(Buffer.from(extendedPanID)),
                 channel,
             };
         });